add utils file
