+++ src/text_clf/main.py
@@ -1,68 +1,20 @@
 import ast
 import datetime
 import json
-import random
 import shutil
 import time
-from argparse import ArgumentParser
 from pathlib import Path
 from typing import Any, Dict, Tuple
 
 import joblib
-import numpy as np
 import pandas as pd
-import yaml
 from sklearn.feature_extraction.text import TfidfVectorizer
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import classification_report
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import LabelEncoder
 
-
-def get_argparse() -> ArgumentParser:
-    """
-    Get argument parser.
-
-    :return: argument parser.
-    :rtype: ArgumentParser
-    """
-
-    parser = ArgumentParser(prog="text-clf")
-    parser.add_argument(
-        "--config",
-        type=str,
-        required=False,
-        default="config.yaml",
-        help="Path to config",
-    )
-
-    return parser
-
-
-def get_config(path_to_config: str) -> Dict[str, Any]:
-    """
-    Get config.
-
-    :param str path_to_config: path to config.
-    :return: config.
-    :rtype: Dict[str, Any]
-    """
-
-    with open(path_to_config, mode="r") as fp:
-        config = yaml.safe_load(fp)
-
-    return config
-
-
-def set_seed(seed: int) -> None:
-    """
-    Set seed for reproducibility.
-
-    :param int seed: seed.
-    """
-
-    random.seed(seed)
-    np.random.seed(seed)
+from .utils import get_argparse, get_config, set_seed
 
 
 def load_data(
